:recycle: Split up buckets schema
